tcmur: Add emulate XCOPY command support
The EXTENDED COPY command provides a means to copy data from one
set of logical units to another set of logical units or to the same
set of logical units. The entity within a SCSI device that manages
the command processing for EXTENDED COPY commands and RECEIVE COPY
RESULTS commands is the copy manager. The copy manager is responsible
for copying data from the source devices to the destination devices.
The copy source and destination devices are logical units that may
reside in different SCSI devices or the same SCSI device. It is
possible that the copy source device, copy destination device, and
the copy manager are the same logical unit.

Currently we are using sync way to do the emulate to simplify the
initial patches, and the async way will be added later as commented
by @liuy.

Signed-off-by: Xiubo Li <lixiubo@cmss.chinamobile.com>